onedrive_agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
tenacity>=8.2.0
msal>=1.24.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
python-json-logger>=2.0.7
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiofiles>=23.2.0
